Remover tabla solo si existe

+++ db/migrate/20160513141707_remove_statuses_table.rb
@@ -1,5 +1,9 @@
 class RemoveStatusesTable < ActiveRecord::Migration[5.0]
-  def change
-    drop_table :statuses
+  def up
+    # Remover tabla obsoleta solo en caso que ya exista
+    drop_table :statuses if ActiveRecord::Base.connection.table_exists? :statuses
+  end
+
+  def down
   end
 end
